requirements.txt
netmiko>=4.3.0
mcp>=1.0.0
pydantic>=2.0.0
asyncio-mqtt>=0.16.0
cryptography>=41.0.0
paramiko>=3.3.0
textfsm>=1.1.3
ntc-templates>=4.0.0
pyyaml>=6.0
click>=8.1.0
